### 2024-03-16 #6496 @editorialbot
[PRE REVIEW]: findSVI: an R package to calculate the Social Vulnerability Index at multiple geographical levels
**Submitting author:** @heli-xu (<a href="http://orcid.org/0000-0002-9792-2727">Heli Xu</a>)
**Repository:** https://github.com/heli-xu/findSVI
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.2
