### 2024-03-16 #6496 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - None
>
> MISSING DOIs
>
> - No DOI given, and none found for title: A social vulnerability index for disaster manageme...
> - No DOI given, and none found for title: CDC/ATSDR SVI data and documentation download
> - No DOI given, and none found for title: tidycensus: Load US Census boundary and attribute ...
> - No DOI given, and none found for title: Measuring community vulnerability to natural and a...
> - 10.1038/s41591-021-01379-6 may be a valid DOI for title: Equitable allocation of COVID-19 vaccines in the U...
> - 10.1016/j.jamcollsurg.2020.11.024 may be a valid DOI for title: High social vulnerability and “textbook outcomes” ...
> - 10.1016/j.surg.2021.06.001 may be a valid DOI for title: Profiles in social vulnerability: the association ...
> - No DOI given, and none found for title: Spatiotemporal distribution of power outages with ...
> - 10.1016/j.ijdrr.2020.101517 may be a valid DOI for title: Spatial exploration of the CDC’s Social Vulnerabil...
> - 10.1016/j.jacadv.2023.100577 may be a valid DOI for title: Impact of social vulnerability and demographics on...
> - 10.1016/j.cpcardiol.2022.101182 may be a valid DOI for title: Neighborhood-level social vulnerability and preval...
> - 10.1161/circulationaha.121.054516 may be a valid DOI for title: Social vulnerability and premature cardiovascular ...
> - 10.1016/j.hjdsi.2021.100611 may be a valid DOI for title: Association of community-level social vulnerabilit...
> - 10.15585/mmwr.mm7012e1 may be a valid DOI for title: County-level COVID-19 vaccination coverage and soc...
> - 10.1016/j.amepre.2020.06.006 may be a valid DOI for title: The impact of social vulnerability on COVID-19 in ...
> - 10.1007/s11606-020-05882-3 may be a valid DOI for title: County-level association of social vulnerability w...
> - No DOI given, and none found for title: Social vulnerability and equity: The disproportion...
> - No DOI given, and none found for title: Spatial inequities in COVID-19 testing, positivity...
> - 10.1136/bmjopen-2020-048086 may be a valid DOI for title: Temporal trends in the association of social vulne...
> - 10.1016/j.lana.2022.100220 may be a valid DOI for title: Racial/ethnic and neighbourhood social vulnerabili...
> - 10.1093/aje/kwac076 may be a valid DOI for title: Heterogeneity in spatial inequities in COVID-19 va...
> - 10.1016/j.socscimed.2022.115307 may be a valid DOI for title: Inequities in spatial accessibility to COVID-19 te...
> - No DOI given, and none found for title: dplyr: A Grammar of Data Manipulation
> - No DOI given, and none found for title: purrr: Functional Programming Tools
> - No DOI given, and none found for title: stringr: Simple, Consistent Wrappers for Common St...
> - No DOI given, and none found for title: tidyr: Tidy Messy Data
> - No DOI given, and none found for title: tidyselect: Select from a Set of Strings
> - No DOI given, and none found for title: rlang: Functions for Base Types and Core R and ’Ti...
> - No DOI given, and none found for title: cli: Helpers for Developing Command Line Interface...
>
> INVALID DOIs
>
> - None
> ```

@heli-xu can you please fix these missing DOIs in your reference source file and notify me here when done?
