### 2024-03-16 83f2625
add doi/url to reference for JOSS submission
