+++ paper/paper.bib
@@ -6,21 +6,23 @@
   number={1},
   pages={0000102202154773551792},
   year={2011},
-  publisher={De Gruyter}
+  publisher={De Gruyter},
+  doi = {https://doi.org/10.2202/1547-7355.1792}
 }
 
 @misc{centers2021cdc,
   title={CDC/ATSDR SVI data and documentation download},
   author={Centers for Disease Control and Prevention/ Agency for Toxic Substances and Disease Registry/ Geospatial Research, Analysis, and Services Program},
   year={2021},
-  url = {https://www.atsdr.cdc.gov/placeandhealth/svi/data_documentation_download.html}
+  url={https://www.atsdr.cdc.gov/placeandhealth/svi/data_documentation_download.html}
 }
 
-@article{walker2022tidycensus,
-  title={tidycensus: Load US Census boundary and attribute data as ‘tidyverse’and ‘sf’-ready dataframes (Version 1.2. 3)},
-  author={Walker, K and Herman, M},
-  journal={Computer Software]. https://CRAN. R-project. org/package= tidycensus},
-  year={2022}
+@Manual{walker2022tidycensus,
+  title = {tidycensus: Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames},
+  author = {Kyle Walker and Matt Herman},
+  year = {2024},
+  note = {R package version 1.6.1},
+  url = {https://walker-data.com/tidycensus/},
 }
 
 @article{flanagan2018measuring,
@@ -31,7 +33,8 @@
   number={10},
   pages={34},
   year={2018},
-  publisher={NIH Public Access}
+  publisher={NIH Public Access},
+  url={https://pubmed.ncbi.nlm.nih.gov/32327766/}
 }
 
 @article{schmidt2021equitable,
@@ -42,7 +45,8 @@
   number={7},
   pages={1298--1307},
   year={2021},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={https://doi.org/10.1038/s41591-021-01379-6}
 }
 
 @article{hyer2021high,
@@ -53,7 +57,8 @@
   number={4},
   pages={351--359},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jamcollsurg.2020.11.024}
 }
 
 @article{paro2021profiles,
@@ -64,7 +69,8 @@
   number={6},
   pages={1777--1784},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.surg.2021.06.001}
 }
 
 @article{do2023spatiotemporal,
@@ -75,7 +81,8 @@
   number={1},
   pages={2470},
   year={2023},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={https://doi.org/10.1038/s41467-023-38084-6}
 }
 
 @article{lehnert2020spatial,
@@ -85,7 +92,8 @@
   volume={46},
   pages={101517},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ijdrr.2020.101517}
 }
 
 @article{ibrahim2023impact,
@@ -96,7 +104,8 @@
   number={7},
   pages={100577},
   year={2023},
-  publisher={American College of Cardiology Foundation Washington DC}
+  publisher={American College of Cardiology Foundation Washington DC},
+  doi={https://doi.org/10.1016/j.jacadv.2023.100577}
 }
 
 @article{bevan2022neighborhood,
[…]
