### 2024-03-17 heli-xu/findSVI#17 @osorensen
Editorial comments on JOSS submission
I am the editor of your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/6496). I will post some suggested edits here before starting the review:

- [ ] In the author affiliations, please include "United States of America", so that it becomes "... Philadelphia, PA, United States of America".
- [ ] The **Summary** section at the beginning of the paper is very well written, but also here please make it clear that the SVI is calculated for the United States of America. For example, in the last sentence in the screenshot below it should be made clear which country you refer to when writing "... and the entire country."

![image](https://github.com/heli-xu/findSVI/assets/21175639/7a63c936-a84e-4d4c-9eea-2a2845bb0a5e)
