### 2024-03-17 heli-xu/findSVI#18 @osorensen
License
The License criteria for [JOSS submissions](https://joss.readthedocs.io/en/latest/review_criteria.html#software-license) state the following:

> There should be an [OSI approved](https://opensource.org/licenses/alphabetical) license included in the repository. Common licenses such as those listed on [choosealicense.com](https://choosealicense.com/) are preferred. Note there should be an actual license file present in the repository not just a reference to the license.

>
> Acceptable: A plain-text LICENSE or COPYING file with the contents of an OSI approved license

>
> Not acceptable: A phrase such as ‘MIT license’ in a README file

To comply with this, please change the name of `LICENSE.md` to simply `LICENSE`, and remove the existing `LICENSE` file.

https://github.com/openjournals/joss-reviews/issues/6496
