### 2024-03-17 #6496 @heli-xu
>@heli-xu can you please fix these missing DOIs in your reference source file and notify me here when done?

Hi, I have updated the DOIs for most of the references, and for one article and several manuals without DOIs I included their URLs (https://github.com/heli-xu/findSVI/commit/83f2625021e8828f35dec358400f44324b100ee8). Please let me know if this works. Thank you for your help!
