### 2024-03-17 adac82d
comply with JOSS license criteria #18
