### 2024-03-19 heli-xu/findSVI#19 @epiben
Harmonise code formatting and contributing section
Dear @heli-xu,

I've been assigned as JOSS reviewer and would like to get started, so I'll be posting comments and suggestions here as per JOSS recommendations. To keep things atomic, I'll raise one thing per issue.

Might I suggest you use automatic styling of the code? Currently, the style is relatively heterogeneous, which makes it more tricky to follow the logic. I'd suggest simply running `styler:::style_active_pkg()`. One might even consider adding this as a pre-commit hook.

Cheers
Ben
