### 2024-03-20 935f5e9
styler::style_pkg() code-files and .Rmd #19
