+++ data-raw/zcta_state_xwalk_2019-21.R
@@ -1,5 +1,5 @@
-#code to generate zcta_state_xwalk2019, zcta_state_xwalk2020, zcta_state_xwalk2021.rda
-#CRAN tidycensus doesn't support pulling zcta data by state (2023-5-23)
+# code to generate zcta_state_xwalk2019, zcta_state_xwalk2020, zcta_state_xwalk2021.rda
+# CRAN tidycensus doesn't support pulling zcta data by state (2023-5-23)
 
 library(readxl)
 library(readr)
@@ -9,39 +9,43 @@ library(stringr)
 
 # zcta_state_xwalk2021 ----------------------------------------------------
 
-#zcta to county relationship file(2020)
+# zcta to county relationship file(2020)
 #
-#zcta_cty20  <- read.table("xls/zcta_county_2020.txt", sep = "|", header = T)
-#only has 9695 rows (ZCTAs), matching to 19 states only
-#use Geocorr2022 data instead:
-#https://mcdc.missouri.edu/applications/geocorr2022.html
+# zcta_cty20  <- read.table("xls/zcta_county_2020.txt", sep = "|", header = T)
+# only has 9695 rows (ZCTAs), matching to 19 states only
+# use Geocorr2022 data instead:
+# https://mcdc.missouri.edu/applications/geocorr2022.html
 zcta_cty20 <- read_csv("xls/zcta_county_2020_geocorr.csv") %>%
-  slice(-1) %>% #a row of col_name explanation
+  slice(-1) %>% # a row of col_name explanation
   clean_names()
 
-##state and abbreviation...
+## state and abbreviation...
 st_abb <- zcta_cty20 %>%
   select(GEOID = county, county_name) %>%
-  mutate(st_code = str_sub(GEOID, 1, 2),
-    st_abb = str_sub(county_name, -2)) %>%
+  mutate(
+    st_code = str_sub(GEOID, 1, 2),
+    st_abb = str_sub(county_name, -2)
+  ) %>%
   select(st_code, st_abb) %>%
   distinct(st_code, .keep_all = TRUE)
 
-#county to state reference file(2021)
-#https://www.census.gov/programs-surveys/popest/geographies/reference-files.html
-cty_st21 <- read_xlsx("xls/all-geocodes-v2021.xlsx", skip = 4) %>% #skip first 4 rows (table titles..)
+# county to state reference file(2021)
+# https://www.census.gov/programs-surveys/popest/geographies/reference-files.html
+cty_st21 <- read_xlsx("xls/all-geocodes-v2021.xlsx", skip = 4) %>% # skip first 4 rows (table titles..)
   clean_names()
 
-#state full name
+# state full name
 add_state <- cty_st21 %>%
   distinct(state_code_fips, .keep_all = TRUE) %>%
-  select(st_code = state_code_fips,
-    state = area_name_including_legal_statistical_area_description) %>%
+  select(
+    st_code = state_code_fips,
+    state = area_name_including_legal_statistical_area_description
+  ) %>%
   left_join(st_abb, by = "st_code") %>%
   drop_na()
 
 
-#add state abb and full name
+# add state abb and full name
 cty_st21_clean <- cty_st21 %>%
   rename(st_code = state_code_fips) %>%
   mutate(GEOID = paste0(st_code, county_code_fips)) %>%
@@ -50,22 +54,25 @@ cty_st21_clean <- cty_st21 %>%
     state,
     GEOID,
     county = area_name_including_legal_statistical_area_description,
-    st_abb)
+    st_abb
+  )
 
 
 zcta_state_xwalk2021 <- zcta_cty20 %>%
   select(GEOID = county, ZCTA = zcta, county_name) %>%
[…]
