+++ R/get_census_data.R
@@ -28,22 +28,24 @@
 #'
 #' @examplesIf Sys.getenv("CENSUS_API_KEY") != ""
 #' # Census API key required
-#'  get_census_data(year = 2018,
+#' get_census_data(
+#'   year = 2018,
 #'   geography = "county",
-#'   state = "PA")
+#'   state = "PA"
+#' )
 #'
 #' @importFrom rlang .data
 #' @export
 
-get_census_data <- function(year,
-  geography,
-  state = NULL,
-  county = NULL,
-  key = NULL,
-  geometry = FALSE,
-  ...)
-{
-  #predicate
+get_census_data <- function(
+    year,
+    geography,
+    state = NULL,
+    county = NULL,
+    key = NULL,
+    geometry = FALSE,
+    ...) {
+  # predicate
   year_valid <- 2012:2021
 
   state_valid <- state_valid
@@ -62,182 +64,174 @@ get_census_data <- function(year,
     cli::cli_abort(c(
       "x" = "`year` contains {length(year)} years.",
       "i" = "Data retrieval is performed one year at a time. To retrieve data (and compute SVI) for multiple years separately, use `find_svi()`."
-      ))
+    ))
   }
 
   if (length(year) == 1) {
-
     if (!(year %in% year_valid)) {
-    cli::cli_abort(c(
-      "x" = "{year} is not a valid input for `year`.",
-      "i" = "Years available for census data retrieval: 2012-2021."
-      ))
-    }
-
-  if (inherits(state, "numeric")) {
-  if (any(!(state %in% state_valid_dbl))) {
-    cli::cli_abort(c(
-      "x" = "One or more elements of {state} is not a valid input for `state`.",
-      "i" = "Use `state_valid` for a table of valid input (besides 'US'). Note state full names/abbreviations are characters; FIPS codes are numbers."
-    ))
-  }
-}
-
-  if(inherits(state, "character")) {
-    if (any(!(state %in% state_valid_chr_us))) {
       cli::cli_abort(c(
-        "x" = "One or more elements of {state} is not a valid input for `state`.",
-        "i" = "Use `state_valid` for a table of valid input (besides 'US'). Note state full names/abbreviations are characters; FIPS codes are numbers."
-      ))
-    }
-  }
-
-#after input validated
-  filename <- paste0("census_variables_", year)
-
-  var_list <- get(filename) %>%
-    unlist() %>%
-    unname()
-
-#state = 1
-if (length(state) == 1) {
-
-  ##US
-  if (state == "US") {
-    raw_data <- tidycensus::get_acs(
-      geography = geography,
-      state = NULL,
-      year = year,
-      variables = var_list,
-      output = "wide",
-      geometry = geometry,
-      ...
-    )
-    return(raw_data)
-  }
-
-  ##zcta >=2019
-  else if(geography == "zcta"&& year >= 2019) {
-    cli::cli_alert_info(
-      "State-specific ZCTA-level data for {year} is currently not supported by Census API.
-Getting nation-based data and selecting ZCTAs in {state}...(it might take a bit longer)"
-      )
-
-  us_data <- tidycensus::get_acs(
-    geography = geography,
-    state = NULL,
-    county = NULL,
-    year = year,
-    variables = var_list,
-    output = "wide",
-    geometry = geometry,
-    ...
-  )
-
-  xwalk_name <- paste0("zcta_state_xwalk", year)
-  st_input <- state
-
-  if (length(county) == 0) {
-  zcta_by_state <- get(xwalk_name) %>%
-    dplyr::filter(.data$state == st_input | .data$st_code == st_input | .data$st_abb == st_input) %>%
-    dplyr::pull("ZCTA")
[…]
