### 2024-03-20 heli-xu/findSVI#19 @heli-xu
Hi Ben,

Thank you for agreeing to review our paper and for your suggestion! We'll plan to add a pre-commit hook for this.

-Heli
