+++ pkgdown.yml
@@ -1,13 +1,13 @@
-pandoc: 2.19.2
-pkgdown: 2.0.7
+pandoc: 3.1.11
+pkgdown: 2.0.8
 pkgdown_sha: ~
 articles:
-  SVI-mapping: SVI-mapping.html
+  findSVI: findSVI.html
   geo_context_of_svi: geo_context_of_svi.html
   svi-covid: svi-covid.html
+  SVI-mapping: SVI-mapping.html
   svi-validation: svi-validation.html
-  findSVI: findSVI.html
-last_built: 2023-09-21T20:23Z
+last_built: 2024-04-18T17:49Z
 urls:
   reference: https://heli-xu.github.io/findSVI/reference
   article: https://heli-xu.github.io/findSVI/articles
