### 2024-04-19 afb4ca2
update column names of cty_cz_2020_xwalk (CZ20 to GEOID2) #16
