### 2024-04-19 6f1dc95
add function get_svi_x() and documentation for svi calculation at customized geographic levels #16
