### 2024-04-22 9feb99b
check() and modify data masking and tidy select binding #16
