+++ R/get_svi.R
@@ -9,7 +9,7 @@
 #'   ranking within a region to indicate the relative social vulnerability of
 #'   the geographic units (communities) in that region.
 #'
-#' @param year The year of interest (available 2014-2021), must match the year
+#' @param year The year of interest (available 2012-2021), must match the year
 #'   specified in retrieving census data.
 #' @param data The census data retrieved by `get_census_data()`.
 #'
