### 2024-05-01 5f1f677
set up new E and EP variables for aggregation
