### 2024-05-03 bc0bbca
add"E" to census var end in cal_table, so that unique() can get rid of repeated var in var_list
