### 2024-05-03 59ebe01
modify to remove random '\t' in extracted var strings
