### 2024-05-03 afcc64f
modify a denominator var in formula (too many NAs)
