+++ R/get_svi_x.R
@@ -34,7 +34,11 @@
 #'
 #' @examplesIf Sys.getenv("CENSUS_API_KEY") != ""
 #' # Census API key required
-#' cty2020 <- get_census_data(year = 2020, geography = "county")
+#' cty2020 <- get_census_data(
+#'     year = 2020,
+#'     geography = "county",
+#'     exp = TRUE
+#'    )
 #'
 #' get_svi_x(year = 2020, data = cty2020, xwalk = cty_cz_2020_xwalk)
 #'
@@ -47,12 +51,16 @@ get_svi_x <- function(year, data, xwalk) {
     cli::cli_abort("The crosswalk does not contain `GEOID` (Census) and `GEOID2 (customized) columns.")
   }
 
-  xwalk_check <- xwalk %>% dplyr::count(GEOID)
+  xwalk_check <- xwalk %>% dplyr::count(.data$GEOID)
   if (!all(xwalk_check$n <= 1)) {
     cli::cli_abort("`GEOID`(Census) level is not completely nested in `GEOID2`(customized) level.")
   }
 
-  ## set up theme 0 vector, because sometimes other E_var calculation refer to them
+  filename <- paste0("variable_cal_exp_", year)
+
+  var_cal_table <- get(filename)
+
+  ## set up theme 0 vector
   var_0 <- var_cal_table %>%
     dplyr::filter(.data$theme == 0)
 
@@ -81,7 +89,6 @@ get_svi_x <- function(year, data, xwalk) {
   names(EP_var_expr) <- EP_var_name
 
 
-
   if ("geometry" %in% colnames(data)) {
     data_tmp <- data %>%
       sf::st_drop_geometry()
@@ -89,11 +96,29 @@ get_svi_x <- function(year, data, xwalk) {
     data_tmp <- data
   }
 
+  ## iterate with E_ vector
+  svi0 <-
+    purrr::map2_dfc(var_0_name, var_0_expr, function(var_0_name, var_0_expr){
+      data_tmp %>%
+        dplyr::transmute(
+          {{var_0_name}} := eval(str2lang({{var_0_expr}}))
+        )
+    }) %>%
+    dplyr::bind_cols(data_tmp, .)
 
+  svi_e <-
+    purrr::map2_dfc(E_var_name, E_var_expr, function(E_var_name, E_var_expr){
+      svi0 %>%
+        dplyr::transmute(
+          {{E_var_name}} := eval(str2lang({{E_var_expr}}))
+        )
+    }) %>%
+    dplyr::bind_cols(svi0, .)
+  #important: keep all retrieved census var (in their original name)
 
 
   if ("NAME" %in% colnames(xwalk)) {
-    data_custom <- data_tmp %>%
+    data_custom <- svi_e %>%
       dplyr::select(-"NAME") %>%
       dplyr::left_join(xwalk, by = "GEOID") %>%
       dplyr::group_by(.data$GEOID2) %>%
@@ -104,7 +129,7 @@ get_svi_x <- function(year, data, xwalk) {
         by = "GEOID2") %>%
       dplyr::rename(GEOID = "GEOID2")
   } else {
-    data_custom <- data_tmp %>%
+    data_custom <- svi_e %>%
       dplyr::select(-"NAME") %>%
       dplyr::left_join(xwalk, by = "GEOID") %>%
       dplyr::group_by(.data$GEOID2) %>%
@@ -113,7 +138,109 @@ get_svi_x <- function(year, data, xwalk) {
       dplyr::mutate(NAME = "")
   }
 
-  cli::cli_alert_success("Finished aggregating census data by customized boundaries.")
+  svi_e_ep <-
[…]
