### 2024-06-02 #6525 @epiben
First off, I'm really sorry for the delay in doing this review; thing have been hectic on my end recently. Generally, I think this is a good submission, but would encourage to do the following before acceptance:
- Enforce the same formatting throughout the source code, e.g., through `styler:::style_active_pkg()`
- Add a section to the README specifying how to contribute
- Extending the test to reach a greater coverage than the current 32%. Output from `covr::package_coverage()`:
```
findSVI Coverage: 32.24%
R/find_svi.R: 0.00%
R/get_census_data.R: 0.00%
R/get_svi.R: 92.91%
```
