### 2024-06-03 092d982
add wrapper function for data retrieval and svi calculation for customized boundaries #16
