+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 export("%>%")
 export(find_svi)
+export(find_svi_x)
 export(get_census_data)
 export(get_svi)
 export(get_svi_x)
