### 2024-06-03 e2962d4
add contributing section to README #19
