### 2024-06-03 heli-xu/findSVI#19 @heli-xu
Linking the JOSS review comment from @epiben for reference: https://github.com/openjournals/joss-reviews/issues/6525#issuecomment-2143985739
