### 2024-06-03 #6525 @heli-xu
Thank you @epiben for reviewing and providing suggestions!
> - Enforce the same formatting throughout the source code, e.g., through styler:::style_active_pkg()
> - Add a section to the README specifying how to contribute

I've modified the code style and added a brief section about contributing in README. https://github.com/heli-xu/findSVI/issues/19

> - Extending the test to reach a greater coverage than the current 32%.

I'm afraid that it would go beyond the `testthat` workflow that we are familiar with, since the data retrieval part of the package relies on census API (using `tidycensus::get_acs()`). The current testing only covers the SVI calculation part (`get_svi()`) using census data stored in the package. We would appreciate any suggestions on testing strategies with API packages, and perhaps we, or someone from the user community, can design a test suite in a future version of the package.

Please let me know if you have further feedback or recommendations. Thanks again!
