### 2024-06-03 #6525 @epiben
That's a valid point. To work around that, I'd probably create your own wrapper around `tidycensus::get_acs()` and call that in your code. So, essentially creating a new file `get_acs_.R` with something like this:

```
#' Wrapper of tidycensus::get_acs() to enable mocking during testing
#' @keywords internal
get_acs_ <- function(...) {
    tidycensus::get_acs(...)
}
```

The, you could replace all your calls to `tidycensus::get_acs` with `get_acs_` and use mocking to pretend you're calling the API during testing: https://testthat.r-lib.org/reference/local_mocked_bindings.html. I know it's a bit of extra work, but I think it would be very beneficial--and a good learning opportunity.
