### 2024-06-07 decf4db
add dataset and documentation for calculation table and var list for 2022 (as per CDC 2022 doc) #20
