+++ R/get_svi.R
@@ -21,18 +21,18 @@
 #'
 #' @examplesIf Sys.getenv("CENSUS_API_KEY") != ""
 #' # Census API key required
-#'  pa2018 <- get_census_data(
-#'     year = 2018,
-#'     geography = "county",
-#'     state = "PA")
+#' pa2018 <- get_census_data(
+#'   year = 2018,
+#'   geography = "county",
+#'   state = "PA"
+#' )
 #'
-#'  get_svi(2018, pa2018)
+#' get_svi(2018, pa2018)
 #'
 #' @importFrom rlang .data
 #'
 #' @export
-get_svi <- function(year, data){
-
+get_svi <- function(year, data) {
   # E_&EP_
 
   filename <- paste0("variable_e_ep_calculation_", year)
@@ -50,8 +50,10 @@ get_svi <- function(year, data){
   ## set up E_ vector
   E_var <-
     var_cal_table %>%
-    dplyr::filter(.data$theme %in% c(1:4),
-      stringr::str_detect(.[[1]], "E_"))
+    dplyr::filter(
+      .data$theme %in% c(1:4),
+      stringr::str_detect(.[[1]], "E_")
+    )
 
   E_var_name <- E_var[[1]]
   E_var_expr <- E_var[[3]]
@@ -60,15 +62,16 @@ get_svi <- function(year, data){
   ## set up EP_ vector
   EP_var <-
     var_cal_table %>%
-    dplyr::filter(.data$theme %in% c(1:4),
-      stringr::str_detect(.[[1]], "EP_"))
+    dplyr::filter(
+      .data$theme %in% c(1:4),
+      stringr::str_detect(.[[1]], "EP_")
+    )
 
   EP_var_name <- EP_var[[1]]
   EP_var_expr <- EP_var[[3]]
   names(EP_var_expr) <- EP_var_name
 
   if ("geometry" %in% colnames(data)) {
-
     data_tmp <- data %>%
       as.data.frame() %>%
       dplyr::select(-"geometry") %>%
@@ -80,63 +83,73 @@ get_svi <- function(year, data){
 
   ## iterate with E_ vector and THEN EP_ vector
   svi0 <-
-    purrr::map2_dfc(var_0_name, var_0_expr, function(var_0_name, var_0_expr){
+    purrr::map2_dfc(var_0_name, var_0_expr, function(var_0_name, var_0_expr) {
       data_tmp %>%
         dplyr::transmute(
-          {{var_0_name}} := eval(str2lang({{var_0_expr}}))
+          {{ var_0_name }} := eval(str2lang({{ var_0_expr }}))
         )
     }) %>%
     dplyr::bind_cols(data_tmp, .)
 
   svi_e <-
-    purrr::map2_dfc(E_var_name, E_var_expr, function(E_var_name, E_var_expr){
+    purrr::map2_dfc(E_var_name, E_var_expr, function(E_var_name, E_var_expr) {
       svi0 %>%
         dplyr::transmute(
-          {{E_var_name}} := eval(str2lang({{E_var_expr}}))
+          {{ E_var_name }} := eval(str2lang({{ E_var_expr }}))
         )
     }) %>%
     dplyr::bind_cols(svi0, .)
 
-
   svi_e_ep <-
-    purrr::map2(EP_var_name, EP_var_expr, function(EP_var_name, EP_var_expr){
+    purrr::map2(EP_var_name, EP_var_expr, function(EP_var_name, EP_var_expr) {
       svi_e %>%
         dplyr::transmute(
-          {{EP_var_name}} := eval(str2lang({{EP_var_expr}}))
+          {{ EP_var_name }} := eval(str2lang({{ EP_var_expr }}))
         )
     }) %>%
[…]
