### 2024-06-13 481aa3f
Add variables for 2022
