### 2024-06-13 37ecda2
Refactor to filter for E_HU >0 and use dplyr::percent_rank()
