+++ R/get_svi.R
@@ -9,7 +9,7 @@
 #'   ranking within a region to indicate the relative social vulnerability of
 #'   the geographic units (communities) in that region.
 #'
-#' @param year The year of interest (available 2014-2021), must match the year
+#' @param year The year of interest (available 2014-2022), must match the year
 #'   specified in retrieving census data.
 #' @param data The census data retrieved by `get_census_data()`.
 #'
@@ -33,13 +33,13 @@
 #'
 #' @export
 get_svi <- function(year, data) {
-  # E_&EP_
+  # E_ & EP_
 
   filename <- paste0("variable_e_ep_calculation_", year)
 
   var_cal_table <- get(filename)
 
-  ## set up theme 0 vector, because sometimes other E_var calculation refer to them
+  ## set up theme 0 vector for E_var calculations
   var_0 <- var_cal_table %>%
     dplyr::filter(.data$theme == 0)
 
@@ -80,7 +80,6 @@ get_svi <- function(year, data) {
     data_tmp <- data
   }
 
-
   ## iterate with E_ vector and THEN EP_ vector
   svi0 <-
     purrr::map2_dfc(var_0_name, var_0_expr, function(var_0_name, var_0_expr) {
@@ -116,7 +115,8 @@ get_svi <- function(year, data) {
       tidyselect::all_of(E_var_name),
       tidyselect::all_of(EP_var_name)
     ) %>%
-    dplyr::mutate(dplyr::across(tidyselect::all_of(EP_var_name), ~ round(.x, 1)),
+    dplyr::mutate(
+      dplyr::across(tidyselect::all_of(EP_var_name), ~ round(.x, 1)),
       E_AGE65 = dplyr::case_when(
         year >= 2017 ~ E_AGE65,
         TRUE ~ round(E_AGE65, 0)
@@ -127,27 +127,30 @@ get_svi <- function(year, data) {
 
   svi_epl <-
     svi_e_ep %>%
-    dplyr::filter(.data$E_TOTPOP > 0) %>% # added according to documentation (removed from ranking, but kept in table)
-    dplyr::select("GEOID", "NAME", tidyselect::all_of(EP_var_name)) %>% # tidyselect, column or external vector
-    tidyr::pivot_longer(!c("GEOID", "NAME"), # all but GEOID and NAME - no need to know total columns
+    dplyr::filter(
+      .data$E_TOTPOP > 0,
+      .data$E_HU > 0
+    ) %>% # remove from ranking, but keep in table
+    dplyr::select(
+      "GEOID",
+      "NAME",
+      tidyselect::all_of(EP_var_name)
+    ) %>%
+    tidyr::pivot_longer(!c("GEOID", "NAME"),
       names_to = "svi_var",
       values_to = "value"
     ) %>%
-    tidyr::drop_na("value") %>% # in case there's *some* variables missing in some tracts
+    tidyr::drop_na("value") %>% # in case of variables missing in tracts
     dplyr::group_by(.data$svi_var) %>%
-    dplyr::mutate(rank = rank(.data$value, ties.method = "min")) %>%
-    # check out count() "wt" arg, if NULL, count rows
-    dplyr::add_count(.data$svi_var) %>%
     dplyr::mutate(
       EPL_var = dplyr::case_when(
-        year >= 2019 ~ (rank - 1) / (n - 1),
-        .data$svi_var == "EP_PCI" ~ 1 - ((rank - 1) / (n - 1)),
-        TRUE ~ (rank - 1) / (n - 1)
[…]
