### 2024-06-13 6274e8c
update documentation for explicit denominator for ep_variables
