### 2024-06-13 30b77e3
add test for 2022 data against cdc svi #20
