### 2024-06-13 4bc5d7f
update documentation, devtools::check() passed
