+++ man/census_variables.Rd
@@ -14,16 +14,17 @@
 \alias{census_variables_2020}
 \alias{census_variables_2021}
 \alias{census_variables_2022}
+\alias{census_variables_exp_2012}
+\alias{census_variables_exp_2013}
+\alias{census_variables_exp_2014}
+\alias{census_variables_exp_2015}
+\alias{census_variables_exp_2016}
+\alias{census_variables_exp_2017}
+\alias{census_variables_exp_2018}
+\alias{census_variables_exp_2019}
 \alias{census_variables_exp_2020}
 \alias{census_variables_exp_2021}
-\alias{census_variables_exp_2019}
-\alias{census_variables_exp_2018}
-\alias{census_variables_exp_2017}
-\alias{census_variables_exp_2016}
-\alias{census_variables_exp_2015}
-\alias{census_variables_exp_2014}
-\alias{census_variables_exp_2013}
-\alias{census_variables_exp_2012}
+\alias{census_variables_exp_2022}
 \title{List of census variables for SVI calculation}
 \format{
 \subsection{a list of census variables, grouped into sublists named t0-t5 or}{
@@ -49,7 +50,7 @@ to \code{variable_e_ep_calculation_} series of tables, where "EP_" (percent)
 variables are retrieved directly from Census when available; Datasets
 starting with \code{census_variables_exp_} contains variables corresponding to
 \code{variable_cal_exp_} tables, where denominators for "EP_" variables are
-explicitly defined using census variables.
+explicitly defined using census variables (adjunct variables are not modified).
 }
 
 An object of class \code{list} of length 6.
@@ -72,25 +73,27 @@ An object of class \code{list} of length 6.
 
 An object of class \code{list} of length 6.
 
-An object of class \code{list} of length 6.
-
-An object of class \code{list} of length 6.
-
-An object of class \code{list} of length 6.
-
-An object of class \code{list} of length 6.
-
-An object of class \code{list} of length 6.
-
-An object of class \code{list} of length 6.
-
-An object of class \code{list} of length 6.
-
-An object of class \code{list} of length 6.
-
-An object of class \code{list} of length 6.
-
 An object of class \code{list} of length 5.
+
+An object of class \code{list} of length 6.
+
+An object of class \code{list} of length 6.
+
+An object of class \code{list} of length 6.
+
+An object of class \code{list} of length 6.
+
+An object of class \code{list} of length 6.
+
+An object of class \code{list} of length 6.
+
+An object of class \code{list} of length 6.
+
+An object of class \code{list} of length 6.
+
+An object of class \code{list} of length 6.
+
+An object of class \code{list} of length 6.
 }
 \source{
 CDC/ATSDR SVI Documentation
@@ -119,25 +122,27 @@ census_variables_2021
 
 census_variables_2022
 
+census_variables_exp_2012
+
+census_variables_exp_2013
+
+census_variables_exp_2014
+
+census_variables_exp_2015
+
+census_variables_exp_2016
+
+census_variables_exp_2017
+
+census_variables_exp_2018
+
+census_variables_exp_2019
+
 census_variables_exp_2020
 
 census_variables_exp_2021
 
-census_variables_exp_2019
-
-census_variables_exp_2018
-
-census_variables_exp_2017
-
-census_variables_exp_2016
-
-census_variables_exp_2015
-
-census_variables_exp_2014
-
-census_variables_exp_2013
-
-census_variables_exp_2012
+census_variables_exp_2022
 }
 \description{
 Each of these datasets contains a list of census variable names for a year
