+++ man/variable_calculation.Rd
@@ -14,16 +14,17 @@
 \alias{variable_e_ep_calculation_2020}
 \alias{variable_e_ep_calculation_2021}
 \alias{variable_e_ep_calculation_2022}
+\alias{variable_cal_exp_2012}
+\alias{variable_cal_exp_2013}
+\alias{variable_cal_exp_2014}
+\alias{variable_cal_exp_2015}
+\alias{variable_cal_exp_2016}
+\alias{variable_cal_exp_2017}
+\alias{variable_cal_exp_2018}
+\alias{variable_cal_exp_2019}
 \alias{variable_cal_exp_2020}
 \alias{variable_cal_exp_2021}
-\alias{variable_cal_exp_2019}
-\alias{variable_cal_exp_2018}
-\alias{variable_cal_exp_2017}
-\alias{variable_cal_exp_2016}
-\alias{variable_cal_exp_2015}
-\alias{variable_cal_exp_2014}
-\alias{variable_cal_exp_2013}
-\alias{variable_cal_exp_2012}
+\alias{variable_cal_exp_2022}
 \title{Table of census variables and formula for SVI calculation}
 \format{
 \subsection{A data frame with 3 columns and various number of rows, depending}{
@@ -63,25 +64,27 @@ An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) w
 
 An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 3 columns.
 
-An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 3 columns.
-
-An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 3 columns.
-
-An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 3 columns.
-
-An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
-
-An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
-
-An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
-
-An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
-
-An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
-
-An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
-
 An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 3 columns.
+
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
+
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
+
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
+
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
+
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
+
+An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 3 columns.
+
[…]
