### 2024-06-14 06e3f7a
modularize a helper function theme_var_df() to generate var list from parsed cal table
