+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: 3.1.11
-pkgdown: 2.0.8
+pkgdown: 2.0.9
 pkgdown_sha: ~
 articles:
   findSVI: findSVI.html
@@ -7,7 +7,7 @@ articles:
   svi-covid: svi-covid.html
   SVI-mapping: SVI-mapping.html
   svi-validation: svi-validation.html
-last_built: 2024-04-18T17:49Z
+last_built: 2024-06-14T22:17Z
 urls:
   reference: https://heli-xu.github.io/findSVI/reference
   article: https://heli-xu.github.io/findSVI/articles
