### 2024-06-24 #6525 @osorensen
👋 @Coral-orca, @zby0327, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
