### 2024-06-25 #6525 @zby0327
Hi @osorensen @heli-xu ,

I have had the opportunity to review your R package and would like to share my thoughts and feedback.
First and foremost, I tested the package with several datasets and found that it provides reasonable results. The documentation is clear and well-written, ensuring that users can easily understand and implement the package's functionalities. The output format of the result variables is intuitive and convenient for viewing the results. However, I found the variable names somewhat unclear, as I am not familiar with SVI-related concepts and need searching to clarify what the variables refer to. It would be beneficial to include explanations for each result variable, either within the documentation or through reference links.
That being said, I do have some concerns regarding the package. Firstly, its functionality is quite limited, catering only to the retrieval of specific website data and calculating a few predefined results (i.e., SVI). The data from the website is restricted to the United States, and if users have data from other countries, they must format it accordingly to use it with the package.
The package’s scope is restricted to calculating SVI, which limits its usability. Additionally, some data is hard-coded within the package, even though it is available on the website. It would be more efficient to integrate the data retrieval functionality directly into the package and eliminate the hard-coded data storage.
Moreover, the dataset is not open-source and requires an API key for access. While it appears that the package can function with limited usage without an API key, there are restrictions. I recommend clarifying the data source's availability and potential ethical issues, given that this is an open-access journal and the data involves human subjects.
Lastly, there is a strong correlation between this package and the tidycensus package, as many functions from tidycensus are called within your package. It may be more beneficial to incorporate your new functionalities directly into the tidycensus package, thereby enhancing its capabilities without the need for a separate package.
Thank you for your attention to these points. I hope my feedback proves helpful for further development of your package and helps the reviewer to do the decision.
