### 2024-06-25 #6525 @heli-xu
Hi @zby0327 , thank you very much for taking the time to review the package and provide feedback! I will try to address your concerns and questions here.
> However, I found the variable names somewhat unclear.

In the documentation of `get_svi()` and `find_svi()`, the link to CDC/SVI documentation is provided as a reference for the description of the variable names.

>Firstly, its functionality is quite limited.

It's true that the functionality of the package is very specific to calculating SVI. However, our main goal is to make it easier to calculate SVI at different geographic levels (especially those that are not available in the CDC SVI database). While traditional workflow involves downloading census data, doing percentile ranks, summing ranks, and doing percentile ranks for the sum, the findSVI workflow can achieve this in one line of code. We and some colleagues have tried both approaches in our research over the past few years, and found the latter much more streamlined and reproducible.

As the SVI is developed by CDC/ATSDR in the US, it is most widely used in studying public health issues in the US, though there have also been efforts to adapt it to other countries. For data from other countries, we do ask users to format it to supply into the function to calculate SVI. However, it might be interesting to incorporate census data retrieval for other countries, like Canada (perhaps via cancensus package), in the future.

>Additionally, some data is hard-coded within the package, even though it is available on the website.

The lists of variables and calculation tables are indeed available on the website in the CDC/SVI documentation, but not in a format that's easily machine-readable. During the development of the package, it took quite some time to extract the information from the pdf tables and check census variables for each year (especially pre-2016 with the previous census data portal). We do agree about keeping the package lightweight, but in this case it's an effective solution for improving performance and facilitating future maintenance.

>I recommend clarifying the data source's availability and potential ethical issues, given that this is an open-access journal and the data involves human subjects.

Census (ACS) data used in the package is anonymized and publicly available; so in this case there is no individual or recorded level data involved. While it is true that an API key is required for using the data retrieval function in the package, any user can request an API key for free.

>Lastly, there is a strong correlation between this package and the tidycensus package.

It's true that the data retrieval function in the package is done via `tidycensus::get_acs()`, but I'm inclined to think that findSVI and tidycensus have very different focus. Tidycensus provides a wonderful interface to obtain a variety of Census data (ACS, Decennial, etc.), whereas findSVI focuses more on processing a set of ACS variables to obtain SVI. There are other examples of R packages that have built additional functionality on top of the data retrieval functionality of tidycensus (e.g. chloroplethr).

Hope this helps to bring in some of our perspectives and rationales. Please let me know if you have further questions/feedback. Thank you again for your time.
