### 2024-06-26 #6525 @zby0327
Hi @heli-xu ,

Thank you very much for your detailed and thoughtful response to my review. I appreciate the clarifications and additional insights you provided.

Your package indeed performs well and achieves its intended functionality. The automation of SVI calculation is a significant improvement over traditional workflows, making the process much more streamlined and reproducible, especially for researchers focused on public health issues in the US.

However, I still have one remaining concern regarding the scope of your package. Specifically, I wonder if providing only the SVI calculation might be somewhat limited. While the package offers a highly efficient method for calculating SVI, I am uncertain whether this meets the JOSS publication standards, particularly the criteria for substantial scholarly effort as outlined here: [JOSS Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort). If that's not the case, I completely agree this package should be accepted by JOSS. @osorensen

Thank you again for your time and consideration. I look forward to your thoughts on this matter.
