### 2024-06-27 #6525 @nilabhrardas
Hi @heli-xu,
I have had the chance to review your package and would like to share my thoughts. Firstly, the functionalities of the package are well-documented, and the vignette is clear and well-written! I was easily able to reproduce the illustrated examples and further test the functions on other datasets.
The package does what it claims - streamlines data retrieval and calculates SVI efficiently with the ability to compute SVIs at different geographical levels. (Personally,) I prefer that the focus of the package is very specific, i.e., to retrieve census data and calculate SVIs, rather than including more functions for retrieving/calculating other related/unrelated 'public health-y' demographics.
The functionality of the functions is restricted to only the US. However, this is not a substantial drawback of the package since other countries and sovereign states are likely to publish data in different standards or even have restricted access to the required data. The reach of the package might benefit from including functionality to transform/transcribe data from other countries (at least for those with unrestricted data availability) for subsequent SVI calculation. While the steps for generating interactive maps using 'leaflet' has been well-documented, the process could be streamlined further by including native visualisation functionalities within the package. This could be an area to be explored for future development.
Overall, I am happy to recommend this paper for publication!
Best wishes.
