### 2024-06-27 #6525 @heli-xu
Hi @nilabhrardas, thank you very much for taking the time to review the package and providing feedback! We're happy to hear that you found the documentation useful and the package effective overall.

>The reach of the package might benefit from including functionality to transform/transcribe data from other countries (at least for those with unrestricted data availability) for subsequent SVI calculation.

That's a good point, and it's true that census data are not always publicly available. As we discussed with @zby0327, it would be an interesting next step to think about incorporating other countries' census data (perhaps leveraging packages like cancensus and geobr).

> the process could be streamlined further by including native visualisation functionalities within the package.

That's definitely another area to explore. It reminds me that we often create functions (via tmap, ggplot2 or leaflet) to make it easier to reproduce maps with different data in our SVI analyses. Perhaps it would be helpful to incorporate some visualisation tools in the package! We'll keep it in mind.

Again, thank you very much for your time and support!
