### 2024-06-29 #6525 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1016/j.jamcollsurg.2020.11.024 is OK
> - 10.1016/j.surg.2021.06.001 is OK
> - 10.1016/j.ijdrr.2020.101517 is OK
> - 10.1016/j.cpcardiol.2022.101182 is OK
> - 10.1016/j.amepre.2020.06.006 is OK
> - 10.1007/s11606-020-05882-3 is OK
> - 10.1111/puar.13264 is OK
> - 10.7326/M20-3936 is OK
> - 10.1136/bmjopen-2020-048086 is OK
> - 10.1016/j.lana.2022.100220 is OK
> - 10.1093/aje/kwac076 is OK
> - 10.1016/j.socscimed.2022.115307 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: CDC/ATSDR SVI data and documentation download
> - No DOI given, and none found for title: tidycensus: Load US Census Boundary and Attribute ...
> - No DOI given, and none found for title: Measuring community vulnerability to natural and a...
> - No DOI given, and none found for title: dplyr: A Grammar of Data Manipulation
> - 10.32614/cran.package.purrr may be a valid DOI for title: purrr: Functional Programming Tools
> - 10.32614/cran.package.stringr may be a valid DOI for title: stringr: Simple, Consistent Wrappers for Common St...
> - 10.32614/cran.package.tidyr may be a valid DOI for title: tidyr: Tidy Messy Data
> - 10.32614/cran.package.tidyselect may be a valid DOI for title: tidyselect: Select from a Set of Strings
> - 10.32614/cran.package.rlang may be a valid DOI for title: rlang: Functions for Base Types and Core R and ’Ti...
> - 10.32614/cran.package.cli may be a valid DOI for title: cli: Helpers for Developing Command Line Interface...
>
> INVALID DOIs
>
> - https://doi.org/10.2202/1547-7355.1792 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1038/s41591-021-01379-6 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1038/s41467-023-38084-6 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.jacadv.2023.100577 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1161/CIRCULATIONAHA.121.054516 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.hjdsi.2021.100611 is INVALID because of 'https://doi.org/' prefix
> - http://dx.doi.org/10.15585/mmwr.mm7012e1 is INVALID because of 'https://doi.org/' prefix
> ```

@heli-xu, could you please go through this list, add missing DOIs and remove 'https://doi.org/' prefixes from your BibTeX source file?
