### 2024-06-30 e884610
added available missing DOIs #8
