+++ paper/paper.bib
@@ -23,6 +23,7 @@
   year = {2024},
   note = {R package version 1.6.1},
   url = {https://walker-data.com/tidycensus/},
+  doi = {10.32614/CRAN.package.tidycensus}
 }
 
 @article{flanagan2018measuring,
@@ -252,6 +253,7 @@
   year = {2023},
   note = {R package version 1.1.4, https://github.com/tidyverse/dplyr},
   url = {https://dplyr.tidyverse.org},
+  doi = {10.32614/CRAN.package.dplyr}
 }
 
 @Manual{purrr,
@@ -260,6 +262,7 @@
   year = {2023},
   note = {R package version 1.0.2, https://github.com/tidyverse/purrr},
   url = {https://purrr.tidyverse.org/},
+  doi = {10.32614/CRAN.package.purrr}
 }
 
 @Manual{stringr,
@@ -268,6 +271,7 @@
   year = {2023},
   note = {R package version 1.5.1, https://github.com/tidyverse/stringr},
   url = {https://stringr.tidyverse.org},
+  doi = {10.32614/CRAN.package.stringr}
 }
 
 @Manual{tidyr,
@@ -276,6 +280,7 @@
   year = {2024},
   note = {R package version 1.3.1, https://github.com/tidyverse/tidyr},
   url = {https://tidyr.tidyverse.org},
+  doi = {10.32614/CRAN.package.tidyr}
 }
 
 @Manual{tidyselect,
@@ -284,6 +289,7 @@
   year = {2024},
   note = {R package version 1.2.1, https://github.com/r-lib/tidyselect},
   url = {https://tidyselect.r-lib.org},
+  doi = {10.32614/CRAN.package.tidyselect}
 }
 
 @Manual{rlang,
@@ -292,6 +298,7 @@
   year = {2024},
   note = {R package version 1.1.3, https://github.com/r-lib/rlang},
   url = {https://rlang.r-lib.org},
+  doi = {10.32614/CRAN.package.rlang}
 }
 
 @Manual{cli,
@@ -300,6 +307,7 @@
   year = {2023},
   note = {R package version 3.6.2, https://github.com/r-lib/cli},
   url = {https://cli.r-lib.org},
+  doi = {10.32614/CRAN.package.cli}
 }
 
 
