### 2024-06-30 fcd57f3
fixed invalid DOI's from 'https://doi.org/' prefix #8
