+++ paper/paper.bib
@@ -7,7 +7,7 @@
   pages={0000102202154773551792},
   year={2011},
   publisher={De Gruyter},
-  doi = {https://doi.org/10.2202/1547-7355.1792}
+  doi = {10.2202/1547-7355.1792}
 }
 
 @misc{centers2021cdc,
@@ -47,7 +47,7 @@
   pages={1298--1307},
   year={2021},
   publisher={Nature Publishing Group US New York},
-  doi={https://doi.org/10.1038/s41591-021-01379-6}
+  doi={10.1038/s41591-021-01379-6}
 }
 
 @article{hyer2021high,
@@ -83,7 +83,7 @@
   pages={2470},
   year={2023},
   publisher={Nature Publishing Group UK London},
-  doi={https://doi.org/10.1038/s41467-023-38084-6}
+  doi={10.1038/s41467-023-38084-6}
 }
 
 @article{lehnert2020spatial,
@@ -106,7 +106,7 @@
   pages={100577},
   year={2023},
   publisher={American College of Cardiology Foundation Washington DC},
-  doi={https://doi.org/10.1016/j.jacadv.2023.100577}
+  doi={10.1016/j.jacadv.2023.100577}
 }
 
 @article{bevan2022neighborhood,
@@ -128,7 +128,7 @@
   pages={1272--1279},
   year={2021},
   publisher={Am Heart Assoc},
-  doi={https://doi.org/10.1161/CIRCULATIONAHA.121.054516}
+  doi={10.1161/CIRCULATIONAHA.121.054516}
 }
 
 @inproceedings{tsai2022association,
@@ -140,7 +140,7 @@
   pages={100611},
   year={2022},
   organization={Elsevier},
-  doi={https://doi.org/10.1016/j.hjdsi.2021.100611}
+  doi={10.1016/j.hjdsi.2021.100611}
 }
 
 @article{hughes2021county,
