### 2024-06-30 7479dbc
fixed invalid DOI's from 'http://dx.doi.org/' prefix #8
