+++ paper/paper.bib
@@ -152,7 +152,7 @@
   pages={431},
   year={2021},
   publisher={Centers for Disease Control and Prevention},
-  doi={http://dx.doi.org/10.15585/mmwr.mm7012e1}
+  doi={10.15585/mmwr.mm7012e1}
 }
 
 @article{karaye2020impact,
