+++ paper/paper.bib
@@ -10,12 +10,6 @@
   doi = {10.2202/1547-7355.1792}
 }
 
-@misc{centers2021cdc,
-  title={CDC/ATSDR SVI data and documentation download},
-  author={Centers for Disease Control and Prevention/ Agency for Toxic Substances and Disease Registry/ Geospatial Research, Analysis, and Services Program},
-  year={2021},
-  url={https://www.atsdr.cdc.gov/placeandhealth/svi/data_documentation_download.html}
-}
 
 @Manual{walker2022tidycensus,
   title = {tidycensus: Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames},
