+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Social Vulnerability Index (SVI) was created by the Centers for Disease Control and Prevention and Agency for Toxic Substances and Disease Registry (CDC/ATSDR) as a tool to assess the resilience of communities in preparation for public health crisis [@flanagan2011social]. In the event of public health emergencies, communities with higher SVI are considered more vulnerable, i.e., more likely to suffer serious consequences such as mortality/physical injuries, displacement and economic loss. The Geospatial Research, Analysis, and Services Program (GRASP) maintains the SVI database and updates the SVI every two years [@centers2021cdc].
+The Social Vulnerability Index (SVI) was created by the Centers for Disease Control and Prevention and Agency for Toxic Substances and Disease Registry (CDC/ATSDR) as a tool to assess the resilience of communities in preparation for public health crisis [@flanagan2011social]. In the event of public health emergencies, communities with higher SVI are considered more vulnerable, i.e., more likely to suffer serious consequences such as mortality/physical injuries, displacement and economic loss. The [Geospatial Research, Analysis, and Services Program (GRASP)](https://www.atsdr.cdc.gov/placeandhealth/svi/data_documentation_download.html) maintains the SVI database and updates the SVI every two years.
 
 The SVI uses 16 variables from the American Community Survey (ACS) [@flanagan2011social].These 16 variables are grouped into 4 themes/domains: Socioeconomic Status, Household Characteristics, Racial & Ethnic Minority Status and Housing Type & Transportation. For each variable, percentile ranks are calculated and aggregated (summed) for each theme and all themes for a geographic level of interest, which are used to generate another set of percentile ranks for a geographic unit as the theme-specific and overall SVIs. The CDC/ATSDR provides the SVI at the census tract and county levels, with percentiles calculated with respect to each state and the entire United States of America.    
 
