### 2024-06-30 bc90218
converted 'flanagan2018measuring' to a hyperlink because it doesn't have a DOI #8
