+++ paper/paper.bib
@@ -20,17 +20,6 @@
   doi = {10.32614/CRAN.package.tidycensus}
 }
 
-@article{flanagan2018measuring,
-  title={Measuring community vulnerability to natural and anthropogenic hazards: the Centers for Disease Control and Prevention’s Social Vulnerability Index},
-  author={Flanagan, Barry E and Hallisey, Elaine J and Adams, Erica and Lavery, Amy},
-  journal={Journal of environmental health},
-  volume={80},
-  number={10},
-  pages={34},
-  year={2018},
-  publisher={NIH Public Access},
-  url={https://pubmed.ncbi.nlm.nih.gov/32327766/}
-}
 
 @article{schmidt2021equitable,
   title={Equitable allocation of COVID-19 vaccines in the United States},
