+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Social Vulnerability Index (SVI) was created by the Centers for Disease Control and Prevention and Agency for Toxic Substances and Disease Registry (CDC/ATSDR) as a tool to assess the resilience of communities in preparation for public health crisis [@flanagan2011social]. In the event of public health emergencies, communities with higher SVI are considered more vulnerable, i.e., more likely to suffer serious consequences such as mortality/physical injuries, displacement and economic loss. The [Geospatial Research, Analysis, and Services Program (GRASP)](https://www.atsdr.cdc.gov/placeandhealth/svi/data_documentation_download.html) maintains the SVI database and updates the SVI every two years.
+The [Social Vulnerability Index (SVI)](https://svi.cdc.gov/Documents/Publications/CDC_ATSDR_SVI_Materials/JEH2018.pdf) was created by the Centers for Disease Control and Prevention and Agency for Toxic Substances and Disease Registry (CDC/ATSDR) as a tool to assess the resilience of communities in preparation for public health crisis [@flanagan2011social]. In the event of public health emergencies, communities with higher SVI are considered more vulnerable, i.e., more likely to suffer serious consequences such as mortality/physical injuries, displacement and economic loss. The [Geospatial Research, Analysis, and Services Program (GRASP)](https://www.atsdr.cdc.gov/placeandhealth/svi/data_documentation_download.html) maintains the SVI database and updates the SVI every two years.
 
 The SVI uses 16 variables from the American Community Survey (ACS) [@flanagan2011social].These 16 variables are grouped into 4 themes/domains: Socioeconomic Status, Household Characteristics, Racial & Ethnic Minority Status and Housing Type & Transportation. For each variable, percentile ranks are calculated and aggregated (summed) for each theme and all themes for a geographic level of interest, which are used to generate another set of percentile ranks for a geographic unit as the theme-specific and overall SVIs. The CDC/ATSDR provides the SVI at the census tract and county levels, with percentiles calculated with respect to each state and the entire United States of America.    
 
@@ -36,7 +36,7 @@ The main aim of the findSVI R package is to extend the application of CDC/ATSDR 
 
 # Statement of Need
 
-The SVI is a widely used indicator to measure the differential impact of public health crisis on communities with different socioeconomic and demographic characteristics. In addition to facilitating effective planning of social services and public assistance for disasters [@flanagan2018measuring; @schmidt2021equitable], the SVI has also provided a valuable metric for characterizing the relationships between social vulnerability and various health outcomes, such as surgical outcomes [@hyer2021high; @paro2021profiles], environmental exposure-related illness [@do2023spatiotemporal; @lehnert2020spatial], cardiovascular diseases [@ibrahim2023impact; @bevan2022neighborhood; @khan2021social], as well as COVID-19 in recent years [@tsai2022association; @hughes2021county; @karaye2020impact; @khazanchi2020county]. In particular, CDC/ATSDR SVI has been used to examine spatial and racial/ethnic inequities in COVID-19 resources and the disproportionate impact of COVID-19 on vulnerable communities [@gaynor2020social; @bilal2021spatial; @islam2021temporal; @bilal2022racial; @bilal2022heterogeneity; @mullachery2022inequities].
[…]
