### 2024-06-30 45eb3a0
Revert "converted 'flanagan2018measuring' to a hyperlink because it doesn't have a DOI #8"
This reverts commit bc902182214017c7c3563561b80568ba554eefac.
