### 2024-06-30 7c8ee5f
replace flanagan2018 without doi (review) with disaster-specific paper #8
