+++ paper/paper.bib
@@ -20,16 +20,16 @@
   doi = {10.32614/CRAN.package.tidycensus}
 }
 
-@article{flanagan2018measuring,
-  title={Measuring community vulnerability to natural and anthropogenic hazards: the Centers for Disease Control and Prevention’s Social Vulnerability Index},
-  author={Flanagan, Barry E and Hallisey, Elaine J and Adams, Erica and Lavery, Amy},
-  journal={Journal of environmental health},
-  volume={80},
-  number={10},
-  pages={34},
-  year={2018},
-  publisher={NIH Public Access},
-  url={https://pubmed.ncbi.nlm.nih.gov/32327766/}
+@article{lue2017mapping,
+  title={Mapping fires and American Red Cross aid using demographic indicators of vulnerability},
+  author={Lue, Evan and Wilson, John P},
+  journal={Disasters},
+  volume={41},
+  number={2},
+  pages={409--426},
+  year={2017},
+  publisher={Wiley Online Library},
+  doi = {10.1111/disa.12198}
 }
 
 @article{schmidt2021equitable,
