+++ paper/paper.md
@@ -36,7 +36,7 @@ The main aim of the findSVI R package is to extend the application of CDC/ATSDR 
 
 # Statement of Need
 
-The SVI is a widely used indicator to measure the differential impact of public health crisis on communities with different socioeconomic and demographic characteristics. In addition to facilitating effective planning of social services and public assistance for disasters [@flanagan2018measuring; @schmidt2021equitable], the SVI has also provided a valuable metric for characterizing the relationships between social vulnerability and various health outcomes, such as surgical outcomes [@hyer2021high; @paro2021profiles], environmental exposure-related illness [@do2023spatiotemporal; @lehnert2020spatial], cardiovascular diseases [@ibrahim2023impact; @bevan2022neighborhood; @khan2021social], as well as COVID-19 in recent years [@tsai2022association; @hughes2021county; @karaye2020impact; @khazanchi2020county]. In particular, CDC/ATSDR SVI has been used to examine spatial and racial/ethnic inequities in COVID-19 resources and the disproportionate impact of COVID-19 on vulnerable communities [@gaynor2020social; @bilal2021spatial; @islam2021temporal; @bilal2022racial; @bilal2022heterogeneity; @mullachery2022inequities].
+The SVI is a widely used indicator to measure the differential impact of public health crisis on communities with different socioeconomic and demographic characteristics. In addition to facilitating effective planning of social services and public assistance for disasters [@lue2017mapping; @schmidt2021equitable], the SVI has also provided a valuable metric for characterizing the relationships between social vulnerability and various health outcomes, such as surgical outcomes [@hyer2021high; @paro2021profiles], environmental exposure-related illness [@do2023spatiotemporal; @lehnert2020spatial], cardiovascular diseases [@ibrahim2023impact; @bevan2022neighborhood; @khan2021social], as well as COVID-19 in recent years [@tsai2022association; @hughes2021county; @karaye2020impact; @khazanchi2020county]. In particular, CDC/ATSDR SVI has been used to examine spatial and racial/ethnic inequities in COVID-19 resources and the disproportionate impact of COVID-19 on vulnerable communities [@gaynor2020social; @bilal2021spatial; @islam2021temporal; @bilal2022racial; @bilal2022heterogeneity; @mullachery2022inequities].
 
 However, many of the applications above had to recalculate the census-tract level SVI to match their units of analysis [@bilal2021spatial; @bilal2022racial; @bilal2022heterogeneity], including ZCTAs, a unit of analysis for which public health data tends to be available quicker in public health emergencies [@bilal2021spatial]. With findSVI, the workflow involved in obtaining SVI can be streamlined and standardized for a specific research question. Provided with year(s), region(s) of reference and a geographic level of interest, findSVI retrieves the required variables from the US Census Bureau data and calculates SVI for communities in the specified area. Furthermore, with its features to retrieve ACS data at more flexible geographic levels and for more recent years, findSVI offers researchers more options for using SVI to accurately reflect the region/year of focus for both association studies and spatial analyses.
 
