### 2024-06-30 #6525 @heli-xu
Hi @osorensen , we've fixed the DOIs (the two references without DOIs were replaced with a hyperlink in text and another relevant paper, respectively. For details https://github.com/heli-xu/findSVI/pull/23). Please let me know if additional changes are needed. Thanks!
