### 2024-07-01 #6525 @ran-codes
@osorensen  will [10.32614/CRAN.package.findSVI](https://doi.org/10.32614/CRAN.package.findSVI) work?

(this is a DOI that was automatically generated via CRAN - the primary R package distirbutor)
