### 2024-07-01 #6525 @osorensen
@ran-codes, the DOI has to be for a repository containing exactly the accepted version. I assume the CRAN DOI will refer to whatever is the most recent version of the package, and hence this is not sufficient. The easiest way is typically to make a Zenodo archive. You can see an example of a recent publication [here](https://zenodo.org/records/12575049).
