+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   svi-covid: svi-covid.html
   SVI-mapping: SVI-mapping.html
   svi-validation: svi-validation.html
-last_built: 2024-06-14T22:17Z
+last_built: 2024-07-01T20:00Z
 urls:
   reference: https://heli-xu.github.io/findSVI/reference
   article: https://heli-xu.github.io/findSVI/articles
