### 2024-07-02 #6525 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
