### 2024-07-02 9716908
add cli message for valid state, year. update 2022 in roxygen
