+++ data-raw/zcta_state_xwalk_2019-21.R
@@ -7,10 +7,106 @@ library(tidyverse)
 library(janitor)
 library(stringr)
 
+# zcta_state_xwalk2022---------------------------------
+
+## zcta-county-----------------------------
+#same as 2020
+zcta_cty20 <- read_csv("xls/zcta_county_2020_geocorr.csv") %>%
+  slice(-1) %>%
+  clean_names()
+
+st_abb <- zcta_cty20 %>%
+  select(GEOID = county, county_name) %>%
+  mutate(st_code = str_sub(GEOID, 1, 2),
+    st_abb = str_sub(county_name, -2)) %>%
+  select(st_code, st_abb) %>%
+  distinct(st_code, .keep_all = TRUE)
+
+
+## county-state--------------------
+cty_st22 <- read_xlsx("xls/all-geocodes-v2022.xlsx", skip = 4) %>%
+  #skip first 4 rows (table titles..)
+  clean_names()
+# change to county-equivalents in CT for 2022
+#https://www.census.gov/programs-surveys/acs/technical-documentation/user-notes/2023-01.html
+#only affect 1 year ACS, 5 year ACS still follows prior2022 boundaries
+
+### modify CT-----------
+cty_st21_ct <- read_xlsx("xls/all-geocodes-v2021.xlsx", skip = 4) %>%
+  clean_names() %>%
+  filter(state_code_fips == "09")
+
+#this double assign just to make column names auto complete
+cty_st21_ct <- cty_st21_ct %>%
+  rename(
+    state_fips_code = state_code_fips,
+    county_fips_code = county_code_fips,
+    county_subdivision_fips_code = county_subdivision_code_fips,
+    place_fips_code = place_code_fips,
+    consolidtated_city_fips_code = consolidtated_city_code_fips,
+    area_name = area_name_including_legal_statistical_area_description
+    )
+
+cty_st22_ct <- cty_st22 %>%
+  filter(!state_fips_code == "09") %>%
+  bind_rows(cty_st21_ct)
+
+#join state name and abbr
+add_state <- cty_st22_ct %>%
+  distinct(state_fips_code, .keep_all = TRUE) %>%
+  select(st_code = state_fips_code,
+    state = area_name) %>% #col names updated
+  left_join(st_abb, by = "st_code") %>%
+  drop_na()
+
+
+#add state abb and full name to county-state
+cty_st22_clean <- cty_st22_ct %>%
+  rename(st_code = state_fips_code) %>%
+  mutate(GEOID = paste0(st_code, county_fips_code)) %>%
+  left_join(add_state, by = "st_code") %>%
+  select(st_code,
+    state,
+    GEOID,
+    county = area_name,
+    st_abb)
+
+cty_st22_clean %>% count(state)
+#53 rows with 1 NA (whole US)
+# turn out the same as 2021
+
+diff <- cty_st22_clean %>%
+  select(GEOID, st_code, county) %>%
+  left_join(cty_st21_clean %>%
+      select(GEOID, st_code, county),
+    by = c("st_code","county")
+) %>%
+  mutate(diff = as.numeric(GEOID.x)-as.numeric(GEOID.y))
+#some of the "city" has two GEOID, like 51000 and 51840
+
+## join zcta-county-state----------
+zcta_state_xwalk2022 <- zcta_cty20 %>%
+  select(GEOID = county, ZCTA = zcta, county_name) %>%
+  #GEOID(county) already character and right length with pad"0"
+  left_join(cty_st22_clean, by = "GEOID", multiple = "all") %>%
+  #there will be multimatch
+  drop_na() %>%
+  select(ZCTA, st_code, county, state, st_abb) %>%
[…]
