+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   svi-covid: svi-covid.html
   SVI-mapping: SVI-mapping.html
   svi-validation: svi-validation.html
-last_built: 2024-07-02T17:42Z
+last_built: 2024-07-02T21:10Z
 urls:
   reference: https://heli-xu.github.io/findSVI/reference
   article: https://heli-xu.github.io/findSVI/articles
