### 2024-07-03 #6525 @samhforbes
Ooh tricky. I'll have to ping @arfon for a fix.
